+++ lib/src/bluetooth/bluetooth_headset.dart
@@ -1,0 +1,11 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import 'package:flutter/services.dart' show MethodChannel;
+
+/// Public API for controlling the Bluetooth Headset Service.
+///
+/// See: https://developer.android.com/reference/android/bluetooth/BluetoothHeadset
+abstract class BluetoothHeadset {
+  static const MethodChannel _channel =
+      MethodChannel('flutter_android/BluetoothHeadset');
+}