+++ lib/src/bluetooth/bluetooth_manager.dart
@@ -1,0 +1,14 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import 'package:flutter/services.dart' show MethodChannel;
+
+import 'bluetooth_adapter.dart' show BluetoothAdapter;
+
+/// High level manager used to obtain an instance of a [BluetoothAdapter] and to
+/// conduct overall Bluetooth management.
+///
+/// See: https://developer.android.com/reference/android/bluetooth/BluetoothManager
+abstract class BluetoothManager {
+  static const MethodChannel _channel =
+      MethodChannel('flutter_android/BluetoothManager');
+}